lmao

+++ src/main/java/com/bluemoonproject/service/UserService.java
@@ -1,12 +1,12 @@
 package com.bluemoonproject.service;
 
-import com.bluemoonproject.constant.PredefinedRole;
-import com.bluemoonproject.dto.request.UserActivateRequest;
+import com.bluemoonproject.enums.PredefinedRole;
 import com.bluemoonproject.dto.request.UserCreationRequest;
 import com.bluemoonproject.dto.request.UserUpdateRequest;
 import com.bluemoonproject.dto.response.UserResponse;
 import com.bluemoonproject.entity.*;
 import com.bluemoonproject.enums.FeeStatus;
+import com.bluemoonproject.enums.ResidencyStatus;
 import com.bluemoonproject.exception.AppException;
 import com.bluemoonproject.exception.ErrorCode;
 import com.bluemoonproject.mapper.UserMapper;
@@ -18,7 +18,6 @@ import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -42,6 +41,7 @@ public class UserService {
     private final OtpService otpService;
     private final GuestRepository guestRepository;
 
+    //method cho admin
     public UserResponse createUser(UserCreationRequest request) {
         log.info("Service: create user");
 
@@ -56,28 +56,30 @@ public class UserService {
         // Convert request to User entity
         User user = userMapper.toUser(request);
 
-
         user.setPassword(passwordEncoder.encode(request.getPassword()));
-
 
         HashSet<Role> roles = new HashSet<>();
         roleRepository.findById(PredefinedRole.USER_ROLE).ifPresent(roles::add);
         user.setRoles(roles);
-
+        user.setResidencyStatus(request.getResidencyStatus());
         // Save user to the database
         user = userRepository.save(user);
 
         return userMapper.toUserResponse(user);
     }
-    
+
+    //admin
     public UserResponse updateUser(Long userId, UserUpdateRequest request) {
-        User user = userRepository.findById(userId).orElseThrow(()->new RuntimeException("User not found"));
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_EXISTS));
         userMapper.updateUser(user, request);
         user.setPassword(passwordEncoder.encode(user.getPassword()));
+        user.setResidencyStatus(request.getResidencyStatus());
 
         return userMapper.toUserResponse(userRepository.save(user));
     }
 
+    // all
     public UserResponse getMyInfo(){
         var context=SecurityContextHolder.getContext();
         String name=context.getAuthentication().getName();
@@ -85,24 +87,29 @@ public class UserService {
         User user=userRepository.findByUsername(name).orElseThrow(()->new AppException(ErrorCode.USER_NOT_EXISTS));
         return userMapper.toUserResponse(user);
     }
+
+    // admin
     public void deleteUser(Long userId){
+        if (!userRepository.existsById(userId)) {
+            throw new AppException(ErrorCode.USER_NOT_EXISTS);
+        }
         userRepository.deleteById(userId);
     }
 
-
+    // admin
     public List<UserResponse> getUsers(){
         log.info("In method get Users");
         return userRepository.findAll().stream().map(userMapper::toUserResponse).toList();
     }
 
-
+    // admin
     public UserResponse getUser(Long id){
         log.info("In method get user by Id");
         return userMapper.toUserResponse(userRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("User not found")));
+                .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_EXISTS)));
     }
 
-
+    // admin
     public User activateAccount(String email) {
         Guest guest = guestRepository.findByEmail(email);
 
@@ -120,7 +127,7 @@ public class UserService {
         HashSet<Role> roles = new HashSet<>();
         roleRepository.findById(PredefinedRole.USER_ROLE).ifPresent(roles::add);
         user.setRoles(roles);
-        user.setRoles(roles);
+        user.setResidencyStatus(ResidencyStatus.THUONG_TRU);
 
         user = userRepository.save(user);
         guestRepository.delete(guest); // Remove guest after activation
@@ -128,6 +135,8 @@ public class UserService {
         return user;
     }
 
+
+    //user
     public List<Fee> getUnpaidFeesForUser() {
         // Find rooms where the user is assigned
         var context=SecurityContextHolder.getContext();
@@ -148,6 +157,7 @@ public class UserService {
         return unpaidFees;
     }
 
+    //user
     public String resetPassword(String email, String otp, String newPassword) {
         if (!otpService.verifyOtp(email, otp)) {
             return "Invalid OTP!";
@@ -164,6 +174,7 @@ public class UserService {
         }
     }
 
+    //admin
     @Transactional
     public void makeUserAdmin(Long userId) {
         User user = userRepository.findById(userId)
@@ -176,12 +187,26 @@ public class UserService {
         userRepository.save(user);
     }
 
+    //user
     public List<Room> getRoomsForUser() {
         var context=SecurityContextHolder.getContext();
         String name=context.getAuthentication().getName();
 
         User user=userRepository.findByUsername(name).orElseThrow(()->new AppException(ErrorCode.USER_NOT_EXISTS));
         return roomRepository.findByUserIdsContaining(user.getId());
+    }
+
+    //admin
+    public List<UserResponse> findUserByResidencyStatus(ResidencyStatus status) {
+        List<User> users = userRepository.findByResidencyStatus(status);
+
+        if (users.isEmpty()) {
+            throw new AppException(ErrorCode.USER_NOT_FOUND);
+        }
+
+        return users.stream()
+                .map(userMapper::toUserResponse)
+                .toList();
     }
 }
 //
